requirements.txt
pymongo==3.10.1
dnspython==1.16.0
metaflow==2.2.3
opencv-python==4.4.0.44
PyYAML==5.3.1
Pillow==8.0.1
